% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capacites_sem_td}
\alias{capacites_sem_td}
\title{Tidy version of part of the \code{capacite} dataset limited to the weekly data \code{capa} per \code{dep} by week \code{sem_*}.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 689 rows and 3 columns.}
\source{
\url{https://github.com/ToulouseDataViz/Hackaviz2020}
}
\usage{
capacites_sem_td
}
\description{
Tidy version of part of the \code{capacite} dataset limited to the weekly data \code{capa} per \code{dep} by week \code{sem_*}.
}
\examples{
capacites_sem_td

# A tibble: 689 x 3
sem    dep      capa
<chr>  <glue>  <dbl>
1 sem_01 dpt_09 110910
2 sem_01 dpt_11 202467

}
\keyword{datasets}
