% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capacites_td}
\alias{capacites_td}
\title{Tidy version of part of the \code{capacite} dataset limited to the not weekly data \code{pop} and \code{hbgt_*} per \code{dep}.}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 13 rows and 6 columns.}
\source{
\url{https://github.com/ToulouseDataViz/Hackaviz2020}
}
\usage{
capacites_td
}
\description{
Total has been removed since it can be computed by summing all the columns.
}
\examples{
capacites_td

# Total hbgt
capacites_td \%>\% mutate(hbgt_total = rowSums(select(., starts_with("hbgt_"))))

}
\keyword{datasets}
