% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsbl.R
\name{capacites_tsbl}
\alias{capacites_tsbl}
\title{\code{capacite_td} dataset as a tsibble.}
\usage{
capacites_tsbl()
}
\value{
a tsibble
}
\description{
Creates and return a tsibble dataset from \code{capacite_td} with the \code{dep} defined as the key.
}
\examples{
capacite_tsbl()

}
