% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comp_departements}
\alias{comp_departements}
\title{Complements / Departements}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 99 rows and 2 columns.}
\source{
\url{https://github.com/ToulouseDataViz/Hackaviz2020}
}
\usage{
comp_departements
}
\description{
Complements / Departements
}
\keyword{datasets}
