% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comp_evenements}
\alias{comp_evenements}
\title{Complements / Evenements}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 202 rows and 5 columns.}
\source{
<https://github.com/ToulouseDataViz/Hackaviz2020>
}
\usage{
comp_evenements
}
\description{
Complements / Evenements
}
\keyword{datasets}
