% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comp_pays}
\alias{comp_pays}
\title{Complements / pays}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13 rows and 2 columns.}
\source{
\url{https://github.com/ToulouseDataViz/Hackaviz2020}
}
\usage{
comp_pays
}
\description{
Complements / pays
}
\keyword{datasets}
