% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nuitees}
\alias{nuitees}
\title{Nuitees.}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 365 rows and 15 columns.}
\source{
<https://github.com/ToulouseDataViz/Hackaviz2020>
}
\usage{
nuitees
}
\description{
Nuitees.
}
\keyword{datasets}
