% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsbl.R
\name{nuitees_tsbl}
\alias{nuitees_tsbl}
\title{\code{nuitees_td} dataset as a tsibble.}
\usage{
nuitees_tsbl()
}
\value{
a tsibble
}
\description{
Creates and return a tsibble dataset from \code{nuitees_td} with the \code{dep} defined as the key.
}
\examples{
nuitees_tsbl()

}
