% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsbl.R
\name{tourists_tsbl}
\alias{tourists_tsbl}
\title{Merge between \code{capacite_td} and \code{nuitees_td} datasets as a tsibble.}
\usage{
tourists_tsbl()
}
\value{
a tsibble
}
\description{
Creates and return a tsibble dataset merge of \code{nuitees_tsbl} and \code{capacites_tsbl}
- \code{dep} defined as the key.
- \code{capa} for capacite
- \code{nuitees} for nuitees
}
\examples{
tourists_tsbl()

}
