% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vacances}
\alias{vacances}
\title{Vacances dataset}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 11688 rows and 5 columns.}
\source{
\url{https://github.com/AntoineAugusti/vacances-scolaires}
}
\usage{
vacances
}
\description{
Vacances dataset
}
\keyword{datasets}
