% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vacances_td}
\alias{vacances_td}
\title{Tidy version of the \code{vacances} dataset.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1095 rows and 3 columns.}
\source{
\url{https://github.com/AntoineAugusti/vacances-scolaires}
}
\usage{
vacances_td
}
\description{
Tidy version of the \code{vacances} dataset.
}
\keyword{datasets}
