% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsbl.R
\name{vacances_tsbl}
\alias{vacances_tsbl}
\title{\code{vacances_td} dataset as a tsibble.}
\usage{
vacances_tsbl()
}
\value{
a tsibble
}
\description{
Creates and return a tsibble dataset from \code{vacances_td} with the \code{zone} defined as the key.
}
\examples{
vacances_tsbl()

}
