% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nuitees_td}
\alias{nuitees_td}
\title{Tidy version of the \code{nuitees} dataset.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4745 rows and 3 columns.}
\source{
\url{https://github.com/ToulouseDataViz/Hackaviz2020}
}
\usage{
nuitees_td
}
\description{
The dataset has now only one variable \code{nuitees} and observations are made by \code{date} and \code{dep}.
In consequence \code{total_occitanie} has been removed since it is a \code{sum} of all the \code{dep} at a \code{date} (see examples).
}
\examples{
nuitees_td

# To get total_occitanie
nuitees_td \%>\%
  group_by(date) \%>\%
  summarise(total_occitanie = sum(nuitees))

}
\keyword{datasets}
