% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{par_origines_td}
\alias{par_origines_td}
\title{Tidy version of the \code{par_origines} dataset.}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 493235 rows and 8 columns.}
\usage{
par_origines_td
}
\description{
Tidy version of the \code{par_origines} dataset.
}
\keyword{datasets}
